provider: remove AzureConfig HostCluster and Location
HostCluster settings will be passed with azure-operator flag. It makes
sense as there is only single operator per host cluster. It will also
make host cluster network preallocation easier.

As for Location we don't have a choice. Because of current peering
limitations we have to have the same location for host and guest
cluster. If needed it will be easier to re-add it than support it.