Don't use embedded type for inheritance (#66)
In Go type embedding cannot be used for inheritance. Interfaces are for
this purpose. When shared data type must be extracted for generic code,
it must have accessible field as embedded types cannot be "casted" like
in more traditional object-oriented design model.

Provider specific guest cluster config must be extracted for use in
provider independent resource implementations. Therefore this field must
be able to be extracted to shared type ClusterGuestConfig.