Add node labels to cluster status
Exposing kubernetes labels assigned for node is useful in many ways when
exposing cluster state to end users. This will be then visible via api to end
users even when they don't have straight access to k8s api.

Main reason exposing these now is to get information which node is master and
which one is worker, but having information on instance type, region and
availability zone the node is present in, is also beneficial.