Add KVM.NodeIndexes field to KVM Status object
Unique [unsigned] integer number / index is required for each node for
different use cases. Right now the most pressing one is for iSCSI initiator
name.

As this is very platform specific at the moment, it is only added for KVM. It
is also added as a separate map distinct from Spec masters/workers and status
nodes because this information is needed before the nodes created and it must
be maintained in the cluster. It should also not affect future work when
refactoring the structuring of worker nodes.