Add cluster availability zones into provider spec
In order to have clean separation for management of cluster's availability
zones, they are defined in the spec as desired state. Intention here is then
that when MachineDeployment controller reconciles a new CR, it ensures that
corresponding Cluster has all required availability zones defined. This also
makes implementation of dynamic AZ / public subnet infrastructure management
easier and more straightforward while having the on-demand aspect of it.