Add generation of CRD storageversion if multiple versions are present (#458)
// +kubebuilder:storageversion marks the given CRD version as the one that
should be used for storage (i.e. persisting in etcd). When there are no other
versions of given CRD, it doesn't generate anything. When new version is added,
one must decided case-by-case which version should be stored in etcd.

This is preparation to introduce `infrastructure/v1alpha3` for Azure
Cluster API types and Node Pools.