Add util functions for working with multiple CRD versions
Introduce two functions that can be used to work with different CRD versions on
different platforms.

Main objective here is to provide functionality for `opsctl` so that AWS can
use one version of Cluster API CRDs while Azure uses different one. This way
one can specify which versions to get and which one to set as storage version.